+++ user_analysis/classify_users.py
@@ -1,3 +1,8 @@
+# SPDX-FileCopyrightText: openmod-tracker contributors
+#
+# SPDX-License-Identifier: MIT
+
+
 """GitHub user classifier.
 
 Modified from https://github.com/dpaolella/open-model-analysis/blob/76201927de5a64ed4b6cd20e197e4b58abe7095b/github_user_analyzer.py