+++ user_analysis/util.py
@@ -1,3 +1,8 @@
+# SPDX-FileCopyrightText: openmod-tracker contributors
+#
+# SPDX-License-Identifier: MIT
+
+
 """Utility functions to support user analysis."""
 
 from pathlib import Path