+++ website/pages/1_👤_Deep_Dive_-_User_Interaction_Analysis.py
@@ -1,3 +1,8 @@
+# SPDX-FileCopyrightText: openmod-tracker contributors
+#
+# SPDX-License-Identifier: MIT
+
+
 """Create Streamlit web app sub-page to visualise tool user data.
 
 (C) Open Energy Transition (OET)
@@ -94,7 +99,7 @@ def user_locations_map(
         .to_frame("Number of Users")
         .reset_index(),
         locations="country",
-        locationmode="ISO-3",
+        locationmode="country names",
         color="Number of Users",
         hover_name="country",
         color_continuous_scale=px.colors.sequential.Viridis,